+++ metadata.rb
@@ -5,11 +5,10 @@ license          'Apache-2.0'
 description      'Wrapper cookbook for setting up Ruby on Rails apps'
 source_url       'https://github.com/nollieheel/app-ror'
 issues_url       'https://github.com/nollieheel/app-ror/issues'
-version          '3.2.0'
+version          '4.0.0'
 
 depends 'git',        '~> 11.1.2'
 depends 'ruby_build', '~> 2.3.0'
-depends 'nodejs',     '~> 10.1.0'
 depends 'yarn',       '~> 0.4.3'
 
 chef_version '>= 17.0'