+++ resources/logrotate.rb
@@ -17,6 +17,7 @@
 # limitations under the License.
 
 cb = 'app_ror'
+
 unified_mode true
 
 property :conf_name, String,