+++ test/cookbooks/test/recipes/manage_sidekiq.rb
@@ -2,7 +2,7 @@
 # Cookbook:: test
 # Recipe:: manage_sidekiq
 #
-# Copyright:: 2022, Earth U
+# Copyright:: 2023, Earth U
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.